DTS: Remove the core hotplug on low battery feature
By default BCL is configured to hotplug cpu4
cpu5, cpu6, and cpu7 when the SOC battery
level threshold (currently 5% [default is 10%])
is reached.

And since we already have CPUQuiet, we don't need
to hotplug the CPU again.

This behaviour is unexpected and can result in
major performance drops.

So we reconfigure BCL to avoid hotplugging any
cores once the SOC battery level threshold
is reached.

All credits goes to Alex.

Change-Id: Icda249e494402379d604b67a29279d7d9a125e5d
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Adek Maulana <adek@techdro.id>